Timer now working
Marten helped me fix the issue with the Interval function that caused browsers to crash. We also tidied the code up and arranged it so it is functional programming, given it uses the same function over and over. Sped up the timer interval so that it goes through colours quicker."